changing the models with better models 1

+++ main.py
@@ -1581,10 +1581,10 @@ if st.session_state.page == "app":
         with col2:
             Age = st.text_input('Age of the Person')
         # Engineered features 
-        Glucose_BMI = int(Glucose) * int(BMI)
-        Age_BMI = int(Age) * int(BMI)
-        BMI_Category = 1 if BMI < 25 else (2 if BMI < 30 else 3)
-        Age_Group = 0 if Age < 30 else (1 if Age < 50 else 20)
+        Glucose_BMI = float(Glucose) * float(BMI)
+        Age_BMI = float(Age) * float(BMI)
+        BMI_Category = 1 if float(BMI) < 25 else (2 if float(BMI) < 30 else 3)
+        Age_Group = 0 if float(Age) < 30 else (1 if float(Age) < 50 else 20)
 
         # Predict Button
